making codebase short
